divide player and dev input sytem
