+++ Assets/ZombieHunter/MovementSystem/MovementSystemsContainer.cs
@@ -10,9 +10,9 @@ namespace ZombieHunter.MovementSystem
         public void AddSystems(EcsSystems ecsSystems)
         {
             ecsSystems
+                .Add(new DevInputSystem())
                 .Add(new BlockJumpSystem())
                 .Add(new PlayerInputSystem.PlayerInputSystem())
-                .Add(new DevInputSystem())
                 .Add(new MovementSystem())
                 .Add(new GroundCheckSystem())
                 .Add(new PlayerJumpSystem());
