+++ Assets/ZombieHunter/PlayerInputSystem/DevInputSystem.cs
@@ -11,65 +11,42 @@ namespace ZombieHunter.PlayerInputSystem
 {
     public class DevInputSystem : IEcsRunSystem, IEcsInitSystem
     { 
+        private float _axisX;
+        private float _axisY;
         private float _moveX;
         private float _moveZ;
 
         private bool _isRotateble = true;
 
         //Inject
+        private readonly EcsFilter<ModelData, MouseLookDirectionData> _mouseLookFilter = null;
         private readonly EcsFilter<Tags.Player, DirectionData, ModelData> _movableFilter = null;
         private readonly EcsFilter<Tags.Player, JumpData> _jumpFilter = null;
         private readonly EcsFilter<Tags.RightPlayerWeapon, WeaponData> _rightWeaponFilter = null;
         private readonly EcsFilter<Tags.LeftPlayerWeapon, WeaponData> _leftWeaponFilter = null;
-        private readonly EcsFilter<ModelData, MouseLookDirectionData> _mouseLookFilter = null;
-        
+
         private readonly EcsStartup.DevelopMode _devMode = null;
-        private readonly PlayerInputConfig _inputConfig = null;
-      
 
         private Quaternion _startTransformRotation;
 
         public void Init()
         {
-            Cursor.lockState = CursorLockMode.Locked;
-            _startTransformRotation = _mouseLookFilter.GetEntity(0).Get<ModelData>().ModelTransform.rotation;
-        }
-        
-        public void Run()
-        {
-            SetDirection();
-
-            foreach (var i in _movableFilter)
-            {
-                ref var directionComponent = ref _movableFilter.Get2(i);
-                ref var direction = ref directionComponent.Direction;
-
-                direction.x = _moveX;
-                direction.z = _moveZ;
-
-                ref var model = ref _movableFilter.Get3(i);
-
-                DirectionModifier(model.ModelTransform);
-            }
-
             if (!_devMode.Value)
             {
                 return;
             }
-            
-            if (Input.GetKeyDown(KeyCode.Space))
+            Cursor.lockState = CursorLockMode.Locked;
+           _startTransformRotation = _mouseLookFilter.GetEntity(0).Get<ModelData>().ModelTransform.rotation;
+        }
+        
+        public void Run()
+        {
+            if (!_devMode.Value)
             {
-                Jump();
+                return;
             }
-            if (Input.GetMouseButtonDown(0))
-            {
-                ShootLeftHand();
-            }
-            
-            if (Input.GetMouseButtonDown(1))
-            {
-                ShootRightHand();
-            }
+            GetMouseInput();
+            GetKeyboardInput();
         }
 
         private void ShootLeftHand()
@@ -100,45 +77,77 @@ namespace ZombieHunter.PlayerInputSystem
                 entity.Get<JumpEvent>();
             }
         }
+
+        private void ClampAxis()
+        {
+            _axisY = Mathf.Clamp(_axisY, -90f, 90f);
+        }
+        private void GetMouseAxis()
+        {
+            foreach (var i in _mouseLookFilter)
+            {
+                ref var model = ref _mouseLookFilter.Get1(i);
+                ref var lookComponent = ref _mouseLookFilter.Get2(i);
+
+                var axisX = lookComponent.Direction.x;
+                var axisY = lookComponent.Direction.y;
+
+                var rotateX = Quaternion.AngleAxis(axisX, Vector3.up * Time.deltaTime * lookComponent.MouseSensitivity);
+                var rotateY = Quaternion.AngleAxis(axisY,
+                    Vector3.right * Time.deltaTime * lookComponent.MouseSensitivity);
+
+                model.ModelTransform.rotation = _startTransformRotation * rotateX;
+                lookComponent.CameraTransform.rotation = model.ModelTransform.rotation * rotateY;
+            }
+            
+            _axisX += Input.GetAxis("Mouse X");
+            _axisY -= Input.GetAxis("Mouse Y");
+        }
+        private void GetMouseActions()
+        {
+            if (Input.GetMouseButtonDown(0))
+            {
+                ShootLeftHand();
+            }
+            
+            if (Input.GetMouseButtonDown(1))
+            {
+                ShootRightHand();
+            }
+        }
+        private void GetKeyboardActions()
+        {
+            if (Input.GetKeyDown(KeyCode.Space))
+            {
+                Jump();
+            }
+        }
         
         private void SetDirection()
         {
-            var axis = OVRInput.Get(OVRInput.RawAxis2D.LThumbstick);
-           _moveX = axis.x;
-           _moveZ = axis.y;
-
-           if (!_devMode.Value)
-           {
-               return;
-           }
-           
-           _moveX = Input.GetAxis("Horizontal");
-           _moveZ = Input.GetAxis("Vertical");
-        }
-
-        private void DirectionModifier(Transform playerTransform)
-        {
-            var axis = OVRInput.Get(OVRInput.RawAxis2D.RThumbstick);
-            if (_isRotateble)
+            foreach (var i in _movableFilter)
             {
-                if (axis.x > _inputConfig.TopPrimaryThumbstickRotateLim)
-                {
-                    playerTransform.Rotate(0f,_inputConfig.RotateModifier,0f);
-                    ModiferTimer();
-                } 
-                else if (axis.x < _inputConfig.LowPrimaryThumbstickRotateLim)
-                {
-                    playerTransform.Rotate(0f,-_inputConfig.RotateModifier,0f);
-                    ModiferTimer();
-                }
+                ref var directionComponent = ref _movableFilter.Get2(i);
+                ref var direction = ref directionComponent.Direction;
+
+                direction.x = _moveX;
+                direction.z = _moveZ;
             }
+            
+            _moveX = Input.GetAxis("Horizontal");
+            _moveZ = Input.GetAxis("Vertical");
         }
 
-        private async void ModiferTimer()
+        private void GetMouseInput()
         {
-            _isRotateble = false;
-            await Task.Delay(_inputConfig.DelayPerRotate);
-            _isRotateble = true;
+            ClampAxis();
+            GetMouseAxis();
+            GetMouseActions();
+        }
+        private void GetKeyboardInput()
+        {
+            SetDirection();
+            GetKeyboardActions();
         }
     }
 }
