--- Assets/ZombieHunter/PlayerInputSystem/MouseLookSystem.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: bfe267c3cf3e44069c4fb1c38b3abaf5
-timeCreated: 1646834711
