+++ UserSettings/EditorUserSettings.asset
@@ -6,34 +6,34 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedScenePath-0:
-      value: 224247031146466102030e2b150b4a1e02120a651f2b233e283a5207e3eb311ce3ea37eee21c2d333432e33d1d3b1371e704001fef
-      flags: 0
-    RecentlyUsedScenePath-1:
-      value: 224247031146466102030e2b150b4a1e02120a651f2b233e283a5214e3ed310be1e238eca92f31352d1b
-      flags: 0
-    RecentlyUsedScenePath-2:
-      value: 224247031146466102030e2b150b4a1e02120a65002d21312e305200e1e53a3df1a811e8ea3f0c3f3c0cea721130082beb
-      flags: 0
-    RecentlyUsedScenePath-3:
       value: 224247031146466102030e2b150b4a1e02120a6501273035202c1327d1f9272ce7ea79cde23730722c0ce6281d
       flags: 0
-    RecentlyUsedScenePath-4:
-      value: 224247031146466102030e2b150b4a1e02120a6501273035202c1327d1f9272ce7ea79cde23730731d07e2334a2b0f36e613
-      flags: 0
-    RecentlyUsedScenePath-5:
+    RecentlyUsedScenePath-1:
       value: 224247031146466102030e2b150b4a1e02120a651b2d273f23661936efef7b1ce7ea39a7f234362820
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-2:
       value: 224247031146466102030e2b150b4a1e02120a651c242729283b343df2f5200bfbf422ecea751b39340da01801330e71e704001fef
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-3:
       value: 224247031146466102030e2b150b4a1e02120a650926233d341e1c25e7d32d2bf6e23ba6c33f32337626ea310b701431fb1e10
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-4:
       value: 224247031146466102030e2b150b4a1e02120a651b2d272022272e2af1f43135ade333e4e8751b39340da1290a371526
       flags: 0
-    RecentlyUsedScenePath-9:
+    RecentlyUsedScenePath-5:
+      value: 224247031146466102030e2b150b4a1e02120a650127332328001323f7f40721f1f333e4a81e3a31364dc233112d0416fc1a1c1fd20e07045f1bf30705e6
+      flags: 0
+    RecentlyUsedScenePath-6:
+      value: 224247031146466102030e2b150b4a1e02120a6501273035202c1327d1f9272ce7ea79cde23730731d07e2334a2b0f36e613
+      flags: 0
+    RecentlyUsedScenePath-7:
+      value: 224247031146466102030e2b150b4a1e02120a65012733232805123ce9d32d2bf6e23ba6c33f3233762fe029173b2d30fd012d0efb04441e1f07e917
+      flags: 0
+    RecentlyUsedScenePath-8:
       value: 224247031146466102030e2b150b4a1e02120a650b292b351e2a183de7ae2136ebf32f
+      flags: 0
+    RecentlyUsedScenePath-9:
+      value: 224247031146466102030e2b150b4a1e02120a651c242729283b343df2f5200bfbf422ecea75183d3407dc3f0130047fa3441c05ff1f13
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
